+++ clinica/pipelines/statistics_surface/surfstat/__init__.py
@@ -1,0 +1,4 @@
+from ._surfstat import clinica_surfstat
+from ._utils import get_t1_freesurfer_custom_file_template
+
+__all__ = ["clinica_surfstat", "get_t1_freesurfer_custom_file_template"]
