+++ clinica/pipelines/statistics_surface/surfstat/models/__init__.py
@@ -1,0 +1,3 @@
+from ._factory import GLMModelType, create_glm_model
+
+__all__ = ["create_glm_model", "GLMModelType"]
