+++ clinica/pipelines/statistics_surface/surfstat/models/results/__init__.py
@@ -1,0 +1,9 @@
+from ._plot import StatisticsResultsPlotter
+from ._serialize import StatisticsResultsSerializer
+from ._statistics import StatisticsResults
+
+__all__ = [
+    "StatisticsResults",
+    "StatisticsResultsSerializer",
+    "StatisticsResultsPlotter",
+]
