+++ test/unittests/pipelines/statistics_surface/test_inputs.py
@@ -8,17 +8,21 @@ CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
 
 
 def test_read_and_check_tsv_file_filenotfound_error(tmpdir):
-    from clinica.pipelines.statistics_surface._inputs import _read_and_check_tsv_file
+    from clinica.pipelines.statistics_surface.surfstat._utils import (
+        read_and_check_tsv_file,
+    )
 
     with pytest.raises(FileNotFoundError, match="File foo.tsv does not exist"):
-        _read_and_check_tsv_file(Path("foo.tsv"))
+        read_and_check_tsv_file(Path("foo.tsv"))
 
 
 @pytest.mark.parametrize(
     "columns", [["foo"], ["foo", "bar"], ["participant_id", "bar"]]
 )
 def test_read_and_check_tsv_file_data_errors(tmpdir, columns):
-    from clinica.pipelines.statistics_surface._inputs import _read_and_check_tsv_file
+    from clinica.pipelines.statistics_surface.surfstat._utils import (
+        read_and_check_tsv_file,
+    )
 
     df = pd.DataFrame(columns=columns)
     df.to_csv(tmpdir / "foo.tsv", sep="\t", index=False)
@@ -26,12 +30,14 @@ def test_read_and_check_tsv_file_data_errors(tmpdir, columns):
         ValueError,
         match=r"The TSV data should have at least two columns: participant_id and session_id",
     ):
-        _read_and_check_tsv_file(tmpdir / "foo.tsv")
+        read_and_check_tsv_file(tmpdir / "foo.tsv")
 
 
 def test_read_and_check_tsv_file():
-    from clinica.pipelines.statistics_surface._inputs import _read_and_check_tsv_file
+    from clinica.pipelines.statistics_surface.surfstat._utils import (
+        read_and_check_tsv_file,
+    )
 
-    df = _read_and_check_tsv_file(Path(CURRENT_DIR) / "data/subjects.tsv")
+    df = read_and_check_tsv_file(Path(CURRENT_DIR) / "data/subjects.tsv")
     assert len(df) == 7
     assert set(df.columns) == {"group", "age", "sex"}
