+++ test/unittests/pipelines/statistics_surface/test_model.py
@@ -7,6 +7,14 @@ import pytest
 from brainstat.stats.terms import FixedEffect
 from numpy.testing import assert_array_almost_equal, assert_array_equal
 
+from clinica.pipelines.statistics_surface.surfstat.models._correlation import (
+    CorrelationGLM,
+)
+from clinica.pipelines.statistics_surface.surfstat.models._group import (
+    GroupGLM,
+    GroupGLMWithInteraction,
+)
+
 CURRENT_DIR = os.path.dirname(os.path.realpath(__file__))
 
 
@@ -16,7 +24,9 @@ def df():
 
 
 def test_missing_column_error(df):
-    from clinica.pipelines.statistics_surface._model import _check_column_in_df
+    from clinica.pipelines.statistics_surface.surfstat.models._utils import (
+        check_column_in_df,
+    )
 
     with pytest.raises(
         ValueError,
@@ -25,18 +35,22 @@ def test_missing_column_error(df):
             "provided TSV file. Please make sure that there is no typo"
         ),
     ):
-        _check_column_in_df(df, "foo")
+        check_column_in_df(df, "foo")
 
 
 def test_is_categorical(df):
-    from clinica.pipelines.statistics_surface._model import _categorical_column
+    from clinica.pipelines.statistics_surface.surfstat.models._utils import (
+        is_categorical,
+    )
 
-    assert _categorical_column(df, "sex")
-    assert not _categorical_column(df, "age")
+    assert is_categorical(df, "sex")
+    assert not is_categorical(df, "age")
 
 
 def test_build_model_term_error(df):
-    from clinica.pipelines.statistics_surface._model import _build_model_term
+    from clinica.pipelines.statistics_surface.surfstat.models._utils import (
+        _build_model_term,
+    )
 
     assert isinstance(_build_model_term("sex", df), FixedEffect)
 
@@ -47,9 +61,9 @@ def test_build_model_intercept(design, df):
     Especially, the fact that adding explicitly the intercept doesn't change the results.
     Test also that spaces in the design expression have no effect.
     """
-    from clinica.pipelines.statistics_surface._model import _build_model
+    from clinica.pipelines.statistics_surface.surfstat.models._utils import build_model
 
-    model = _build_model(design, df)
+    model = build_model(design, df)
     assert isinstance(model, FixedEffect)
     assert len(model.m.columns) == 2
     assert_array_equal(model.intercept, np.array([1, 1, 1, 1, 1, 1, 1]))
@@ -57,16 +71,16 @@ def test_build_model_intercept(design, df):
 
 
 def test_build_model(df):
-    from clinica.pipelines.statistics_surface._model import _build_model
+    from clinica.pipelines.statistics_surface.surfstat.models._utils import build_model
 
-    model = _build_model("1 + age + sex", df)
+    model = build_model("1 + age + sex", df)
     assert isinstance(model, FixedEffect)
     assert len(model.m.columns) == 4
     assert_array_equal(model.intercept, np.array([1, 1, 1, 1, 1, 1, 1]))
     assert_array_equal(model.age, np.array([78.0, 73.4, 70.8, 82.3, 60.6, 72.1, 74.2]))
     assert_array_equal(model.sex_Female, np.array([1, 0, 1, 1, 1, 0, 1]))
     assert_array_equal(model.sex_Male, np.array([0, 1, 0, 0, 0, 1, 0]))
-    model = _build_model("1 + age + sex + age * sex", df)
+    model = build_model("1 + age + sex + age * sex", df)
     assert isinstance(model, FixedEffect)
     assert len(model.m.columns) == 6
     assert_array_equal(model.intercept, np.array([1, 1, 1, 1, 1, 1, 1]))
@@ -82,6 +96,22 @@ def test_build_model(df):
     )
 
 
+def test_base_glm_instantiation_error(df):
+    """Test that the base abstract GLM class cannot be instantiated."""
+    from clinica.pipelines.statistics_surface.surfstat.models._base import GLM
+
+    with pytest.raises(NotImplementedError):
+        GLM("1 + age", df, "feature_label", "age")
+
+
+@pytest.mark.parametrize(
+    "model,contrast",
+    [
+        (CorrelationGLM, "age"),
+        (GroupGLM, "sex"),
+        (GroupGLMWithInteraction, "age * sex"),
+    ],
+)
 @pytest.mark.parametrize(
     "parameters",
     [
@@ -94,51 +124,48 @@ def test_build_model(df):
         },
     ],
 )
-def test_glm_instantiation(df, parameters):
-    from clinica.pipelines.statistics_surface._model import GLM
+def test_common_parameters_glm_instantiation(df, model, contrast, parameters):
+    model_instance = model("1 + age", df, "feature_label", contrast, "group")
 
-    model = GLM("1 + age", df, "feature_label", "age")
-    assert not model._two_tailed
-    assert model._correction == ["fdr", "rft"]
-    assert model.feature_label == "feature_label"
-    assert model.fwhm == 20
-    assert model.threshold_uncorrected_pvalue == 0.001
-    assert model.threshold_corrected_pvalue == 0.05
-    assert model.cluster_threshold == 0.001
-    assert model.contrasts == dict()
-    assert model.filenames == dict()
-    assert model.contrast_names == list()
-    assert isinstance(model.model, FixedEffect)
+    assert not model_instance._two_tailed
+    assert model_instance._correction == ["fdr", "rft"]
+    assert model_instance.feature_label == "feature_label"
+    assert model_instance.fwhm == 20
+    assert model_instance.threshold_uncorrected_pvalue == 0.001
+    assert model_instance.threshold_corrected_pvalue == 0.05
+    assert model_instance.cluster_threshold == 0.001
+    assert model_instance.filenames == dict()
+    assert isinstance(model_instance.model, FixedEffect)
 
 
 @pytest.mark.parametrize("contrast", ["age", "-age"])
 def test_correlation_glm_instantiation(df, contrast):
-    from clinica.pipelines.statistics_surface._model import CorrelationGLM
+    from clinica.pipelines.statistics_surface.surfstat.models._contrast import (
+        CorrelationContrast,
+    )
 
-    model = CorrelationGLM("1 + age", df, "feature_label", contrast, "group_label")
+    model = CorrelationGLM("1 + age", df, "feature_label", contrast, "group")
     assert not model.with_interaction
-    assert model.group_label == "group_label"
+    assert model.group_label == "group"
     assert model.feature_label == "feature_label"
     assert model.fwhm == 20
     sign = "positive" if contrast == "age" else "negative"
-    assert model.contrast_sign == sign
-    assert model.absolute_contrast_name == "age"
-    assert isinstance(model.contrasts, dict)
     assert len(model.contrasts) == 1
+    assert isinstance(model.contrasts[0], CorrelationContrast)
+    assert model.contrasts[0].sign == sign
+    assert model.contrasts[0].name == contrast
     mult = 1 if sign == "positive" else -1
     assert_array_equal(
-        model.contrasts[contrast].values,
+        model.contrasts[0].built_contrast.values,
         mult * np.array([78.0, 73.4, 70.8, 82.3, 60.6, 72.1, 74.2]),
     )
     with pytest.raises(ValueError, match="Unknown contrast foo"):
-        model.filename_root("foo")
-    expected = f"group-group_label_correlation-age-{sign}_measure-feature_label_fwhm-20"
-    assert model.filename_root(contrast) == expected
+        model.get_output_filename("foo")
+    expected = f"group-group_correlation-age-{sign}_measure-feature_label_fwhm-20"
+    assert model.get_output_filename(contrast) == expected
 
 
 def test_group_glm_instantiation(df):
-    from clinica.pipelines.statistics_surface._model import GroupGLM
-
     with pytest.raises(
         ValueError,
         match="Contrast should refer to a categorical variable for group comparison.",
@@ -148,24 +175,22 @@ def test_group_glm_instantiation(df):
     assert not model.with_interaction
     assert model.group_label == "group_label"
     assert model.fwhm == 20
-    assert isinstance(model.contrasts, dict)
+    assert isinstance(model.contrasts, list)
     contrast_names = ["Female-lt-Male", "Male-lt-Female"]
-    assert set(model.contrasts.keys()) == set(contrast_names)
+    assert set(model.contrast_names) == set(contrast_names)
     for contrast_name, sign in zip(contrast_names, [-1, 1]):
         assert_array_equal(
-            model.contrasts[contrast_name].values,
+            model.get_contrast_by_name(contrast_name).built_contrast,
             sign * np.array([1, -1, 1, 1, 1, -1, 1]),
         )
     with pytest.raises(ValueError, match="Unknown contrast foo"):
-        model.filename_root("foo")
+        model.get_output_filename("foo")
     for contrast_name in contrast_names:
         expected = f"group-group_label_{contrast_name}_measure-feature_label_fwhm-20"
-        assert model.filename_root(contrast_name) == expected
+        assert model.get_output_filename(contrast_name) == expected
 
 
 def test_group_glm_with_interaction_instantiation(df):
-    from clinica.pipelines.statistics_surface._model import GroupGLMWithInteraction
-
     with pytest.raises(
         ValueError,
         match=(
@@ -173,34 +198,29 @@ def test_group_glm_with_interaction_instantiation(df):
             "variable and one categorical variable."
         ),
     ):
-        GroupGLMWithInteraction("1 + age", df, "feature_label", "age", "group_label")
+        GroupGLMWithInteraction("1 + age", df, "feature_label", "age", "group")
     model = GroupGLMWithInteraction(
-        "1 + age", df, "feature_label", "age * sex", "group_label"
+        "1 + age", df, "feature_label", "age * sex", "group"
     )
     assert model.with_interaction
-    assert model.group_label == "group_label"
+    assert model.group_label == "group"
     assert model.fwhm == 20
-    assert isinstance(model.contrasts, dict)
+    assert isinstance(model.contrasts, list)
     assert len(model.contrasts) == 1
     assert_array_equal(
-        model.contrasts["age * sex"].values,
+        model.get_contrast_by_name("age * sex").built_contrast,
         np.array([78.0, -73.4, 70.8, 82.3, 60.6, -72.1, 74.2]),
     )
     with pytest.raises(ValueError, match="Unknown contrast foo"):
-        model.filename_root("foo")
+        model.get_output_filename("foo")
     assert (
-        model.filename_root("age * sex")
+        model.get_output_filename("age * sex")
         == "interaction-age * sex_measure-feature_label_fwhm-20"
     )
 
 
 def test_create_glm_model(df):
-    from clinica.pipelines.statistics_surface._model import (
-        CorrelationGLM,
-        GroupGLM,
-        GroupGLMWithInteraction,
-        create_glm_model,
-    )
+    from clinica.pipelines.statistics_surface.surfstat.models import create_glm_model
 
     model = create_glm_model(
         "correlation", "age", df, "age", feature_label="feature_label"
@@ -227,7 +247,9 @@ def test_create_glm_model(df):
 
 
 def test_p_value_results():
-    from clinica.pipelines.statistics_surface._model import PValueResults
+    from clinica.pipelines.statistics_surface.surfstat.models.results._statistics import (
+        PValueResults,
+    )
 
     pvalues = np.random.random((5, 10))
     threshold = 0.3
@@ -248,29 +270,34 @@ def test_statistics_results_serializer(tmp_path):
 
     from scipy.io import loadmat
 
-    from clinica.pipelines.statistics_surface._model import (
-        CorrectedPValueResults,
-        PValueResults,
+    from clinica.pipelines.statistics_surface.surfstat.models.results import (
         StatisticsResults,
         StatisticsResultsSerializer,
+    )
+    from clinica.pipelines.statistics_surface.surfstat.models.results._statistics import (
+        CorrectedPValueResults,
+        PValueResults,
     )
 
     dummy_input = np.empty([3, 6])
     uncorrected = PValueResults(*[dummy_input] * 2, 0.1)
     corrected = CorrectedPValueResults(*[dummy_input] * 3, 0.2)
     results = StatisticsResults(*[dummy_input] * 2, uncorrected, dummy_input, corrected)
-    serializer = StatisticsResultsSerializer(str(tmp_path / Path("out/dummy")))
-    assert serializer.output_file == str(tmp_path / Path("out/dummy"))
+
+    out_file = tmp_path / "out" / "dummy"
+    serializer = StatisticsResultsSerializer(out_file)
+    assert serializer.output_file == out_file
     assert serializer.json_extension == "_results.json"
     assert serializer.json_indent == 4
     assert serializer.mat_extension == ".mat"
+
     with pytest.raises(
         NotImplementedError, match="Serializing method foo is not implemented."
     ):
         serializer.save(results, "foo")
     serializer.save(results, "json")
-    assert os.path.exists(tmp_path / Path("out/dummy_results.json"))
-    with open(tmp_path / Path("out/dummy_results.json"), "r") as fp:
+    assert (tmp_path / "out" / "dummy_results.json").exists()
+    with open(tmp_path / "out" / "dummy_results.json", "r") as fp:
         serialized = json.load(fp)
     serializer.save(results, "mat")
     names = [
@@ -288,9 +315,9 @@ def test_statistics_results_serializer(tmp_path):
         "FDR",
     ]
     for name, key in zip(names, keys):
-        assert os.path.exists(tmp_path / Path(f"out/dummy_{name}.mat"))
-        mat = loadmat(tmp_path / Path(f"out/dummy_{name}.mat"))
-        if key in ["uncorrectedpvaluesstruct", "correctedpvaluesstruct"]:
+        assert (tmp_path / "out" / f"dummy_{name}.mat").exists()
+        mat = loadmat(tmp_path / "out" / f"dummy_{name}.mat")
+        if key in ("uncorrectedpvaluesstruct", "correctedpvaluesstruct"):
             assert_array_almost_equal(mat[key]["P"][0, 0], dummy_input)
             assert_array_almost_equal(mat[key]["mask"][0, 0], dummy_input)
         else:
